states: Fix a bug in merging states
What's going on here is that when when we're merging states then we have
to set the cur_stree, but we were only setting a partial stree.  So you
have a bunch of states, then you come to a compound condition:

	if (a && b && c) {

We start with a full slist, then we create a fake stree on top, then we
merge the unmatched states and we want to look at the stree but we only
had the partial fake stree, not the initial full stree.  This made a
difference to some new code in smatch_param_used.c where I want to see
if a parameter was set.  That information wasn't in the partial stree.

So we have to use use a fake stree which gives all the states, then we set
the ones which were set on one side of the condition.

(This changelog makes a lot of sense if you already know how fake strees
work).

I could have left the read_only code but it wasn't important and it just
made this patch more complicated for no reason so I deleted it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>