param_set: improve param_was_set_var_sym()
This code does a huge loop and that seems like it would be terribly slow
so I changed it to a smaller loop.

Then I noticed that it didn't handle unmatched states, in other words,
it returned true if the parameter was set on *any* path, but I feel like it
should return true if the parameter was set on *every* path.  I definitely
need that behavior sometimes, and I can't think why I would need the
original behavior.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>