slist: re-work things again (faster?)
I'm still not totally satisfied with how this is done, but this new way
feels more intuitive perhaps...

The easiest change to explain is smatch_returns.c.  Originally
__set_fake_cur_stree() used to just swap the stree and if I remember
right __swap_cur_stree() was unused.  In smatch_returns.c we do just want
to swap the stree so that's obvious.

When we're merging states then we want to have access to the whole stree
but with our specific states overlayed on top.  Before I was creating
a fake stree and then calling __set_sm() to set the states.  Now I just
have introduced an overlay stree and we look up the states there first.
The rules for that are a bit complicated because it can't nest, and the
strees are marked readonly so we can't set states.  But it should be fast,
I hope.

For the pre-merge hooks I create a fake stree to catch any calls to
set_state() so that those are preserved.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>