db/param_set: record PARAM_SET even when there are too many other states
Sometimes there is just too much information in return_states and so we
can't record anything.  Before we tried to mark things as lost and move on.
But really, I'd like to record PARAM_SET information because that's very
important.

Even when we're just recording PARAM_SET, there can be thousands of states
because we have a whole lot of nested structs which are set.  In that case
I've limited it to three layers deep of nested structs.  Ideally there
would be a better way to just mark things as lost, which would set the
extra state to unknown but without calling the modification hooks.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>