pre-process: fix endless loop in guessing includes
Smatch has a "feature" where it can't find the include where it's supposed
to be then it does a search and include a best guess include.  At first
what happened was the filehandles were leaked and it crashed, but then I
fixed that so it just looped forever.  So then the second part of the fix
was to add a counter in use_best_guess_header_file() so that if we have to
guess over a thousand times then just give up.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>