smdb.py: add find_tagged and parse_warns_tagged commands
The find_tagged command follows a given parameter up the call
stack as far as it can, so long as the parameter contains user
data and the top byte is non-zero. This is helpful to identify
functions that should untag a tagged address.

The parse_warns_tagged command parses the smatch_warns.txt file
and calls find_tagged for each tagged related warning, thus
providing a summary of all issues found and their potential
causes.

Signed-off-by: Andrew Murray <andrew.murray@arm.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>