gen_dma_funcs: add dma_map_single() as a function to check
The address we pass to dma_map_single() function has to be DMAable memory.
Kees is going to add this check to the code so it will generate a runtime
error as well.  Generally this bug is caught at runtime these days so we
seldom see the static checker warnings.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>