"graph" segfaults on top-level asm
The "graph" binary segfaults on this input:
	asm("");
with gdb saying (edited for clarity):
	Program received signal SIGSEGV, Segmentation fault.
		in graph_ep (ep=0x7ffff7f62010) at graph.c:52
	(gdb) p ep->entry
	$1 = (struct instruction *) 0x0

Sadly, the commit that introduced this crash:
	15fa4d60e ("topasm: top-level asm is special")
was (part of a bigger series) meant to fix crashes because
of such toplevel asm statements.

Toplevel ASM statements are quite abnormal:
* they are toplevel but anonymous symbols
* they should be limited to basic ASM syntax but are not
* they are given the type SYM_FN but are not functions
* there is nothing to evaluate or expand about it.
These cause quite a few problems including crashes, even
before the above commit.

So, before handling them more correctly and instead of
adding a bunch of special cases here and there, temporarily
take the more radical approach of stopping to add them to
the list of toplevel symbols.

Fixes: 15fa4d60ebba3025495bb34f0718764336d3dfe0
Reported-by: Vegard Nossum <vegard.nossum@gmail.com>
Analyzed-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>