function_hooks: Use the intersection of fake assigns and known rl returns
I imagine that this mostly affects place where we want to manually
set the return range using smatch_data/db/kernel.return_fixes.  But you
also have a parameter assign.  Right now that's only:

nla_len 0-65531[$0->nla_len - 4]

The problem is that in the original code we first say that the return
is 0-65531, then we say that it's "$0->nla_len - 4" so it can be -4.  We
should instead take the intersection.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>