hooks: clean up the ->data_type field
This code was mixing enum data_type and owners.  Then the data_types array
was declared as a hook_type.

We don't need to store "container->data_type = data_types[type];" because
we can just look up data_types[type] in the __pass_to_client() function.
Then ->data_type is only used to store the ->owner so we can rename it to
that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>