check_deref/deref_check: warn about likely/unlikely() conditions as well
The back story is that these check for inconsistent NULL checking ignore
macros.  Macros have all kinds of unnecessary checks inside them so if
we looked at that, it would lead to tons of false positives.  This causes
a problem though because we have conditions like "if (unlikely(!foo)) {"
and unlikely() is a macro so it's ignored.

So my fix is to make a special case for likely/unlikely().  I think this
should work pretty well for the Linux kernel at least.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>