extra: limit get_other_name_sym_helper() some more
I tried to make a test case for this and I didn't succeed...  I need a
bunch of test cases actually.  The problem was parsing:

static int adpt_i2o_install_device(adpt_hba* pHba, struct i2o_device *d)
{
	d->controller = pHba;
	return 0;
}
	...
	d->controller = pHba;
	adpt_i2o_install_device();

It returns that "d->controller->host is set to pHba->host".  Which is true.
Ideally, just recording that "d->controller = pHba;" would be enough, but
I don't have to code written to do that, and the information is correct and
correct information should not lead to a bug.

But then when we parse the returns to set d->controller->host to
0,4096-ptr_max, the get_other_name() says we need to set "pHba->host" as
well.  Which we don't.

In this case we are setting it to it's original values so the values are
not wrong but it was modified that means it over writes all the
implications.

I'm not totally sure this fix is correct, but I'm committing it anyway
because I want to record the rational before I forget.  Again, I need to
write some test code for get_other_name() because I really don't understand
all the rules very well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>