rosenberg: initialization and struct assignments don't fill holes
Before when we had an initializer:

	struct foo bar = {};

The struct holes were marked as cleared but sometimes they are not.  The
other thing is that when you have an assignment like:

	foo = bar;

where foo and bar are structs, then sometimes it translates that into a
series of assignments like:

	foo.a = bar.a;
	foo.b = bar.b;

Instead of just doing a memcpy() which clears all the holes.  (I'm not
using the right words, but hopefully this commit message is clear).

I had known this was an issue and I had thought about turning this on
before but I was worried it would cause a lot of false positives.  Now it
seems bearable.  I don't know if that's because of other improvements in
Smatch or because my standards have gotten lower.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>