remove unneeded MOD_TYPE
MOD_TYPE is used for the sparse extension which allow
to directly compare types with each others.

Expressions for direct type are EXPR_TYPE with the type in
expr->symbol and the expression itself having it's type
(expr->ctype) set to &type_ctype. This is one of the few
base/builtin types and is the only one which can have MOD_TYPE.

However, a specific modifier is not needed, the address of
the symbol can simple be used (like it is done for 'bad_ctype'
or 'incomplete_ctype'). Also, there is only a single place where
MOD_TYPE is tested: is_type_ctype(), itself used a single time.

So:
* rewrite the unique test using is_type_ctype() by directly
  comparing with &type_ctype instead;
* remove the now unused is_type_ctype();
* remove MOD_TYPE from type_ctype's definition;
* remove MOD_TYPE's definition;
and spare one precious bit for other modifiers.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>