math: for inline functions let's tree sval parameters as constant
Smatch treats constants and known values slightly differently...  I guess
the reason behind that is a lack of confidence.  This lack of confidence is
probably justified because Smatch doesn't understand loops very well and
generally parses a lot of edge cases incorrectly.

Say we have code looks like this:

	frob(0);

In the caller we know that 0 is a constant, but in the function we don't.

Now, I have changed it around to say that if we're parsing an inline
function and we have a parameter which is known then let's treat it as a
constant.  My hope is that this will fix more false positives than it
introduces.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>