buf_size: improve how variable buffer sizes are stored
Before we were recording returned buffer sizes in two places.  For known
buffers it was done in smatch_buf_size.c and for variable buffer sizes it
was recorded in smatch_parse_call_math.c.

But now we want to return a range of potential buffer sizes like 10-200
and maybe also that it's equal to "$0 + 10".  So then one of those erases
the other.  We should instead store the information in one go like
"10-200[$0 - 10]".  This approach has the advantage that we can use
call_results_to_rl() to parse the results.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>