function_hooks: don't set the extra state twice
To be honest, I have forgotten what problem this solves.  I think it was
possibly supposed to be part of commit 575aa666c95f ("function_hooks: Use
the intersection of fake assigns and known rl returns").  It probably
doesn't solve any problem but is just intended as a correctness issue where
it doesn't make sense to set the same state twice.

I could imagine that I was confused by the earlier behavior when I was
trying to debug a problem.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>