check_free_strict: merged NULL and freed pointers are freed
There are sometimes where we want to ask if the parent is freed or NULL.
For example we could have a lock:

	free(foo);
	return ret;

In that code we don't want to warn that foo->lock wasn't unlocked because
it was freed so we don't care.

The problem is that some code only frees pointer which are non-NULL so then
when we merge the freed and NULL pointer the result is &merged.  We could
do something complicated where we assume non-NULL and then check whether
the state is &freed, but it's easier to say that the merge between a freed
and a NULL pointer results in a freed pointer.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>