function_ptr: add a huge hack for arrays of pointers
This is really ugly...  Perhaps the ugliest thing I have written in years.

The difference between is_array() and xxx_is_array() is this line:

+	if (type->type != SYM_ARRAY && type->type != SYM_PTR)

I the original only SYM_ARRAY counts as arrays but here pointers are
counted.  If we don't do this a bunch of array assignments are missed
and pointer math is missed.  I'm not immediately sure how to fix this in
the correct way.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>