db/copy_function_pointers:  improve pointer call tracking
If you have a function pointer which points to a function pointer which
points to a third function pointer and so on then these weren't searched
properly.  What this does is that it recursively copies them so that in
smatch_db.c we can just select the functions without having to follow the
pointer to pointer trail.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>