comparison: improve pre merge hook
In the original code, this would overwrite our state with whatever
smatch_extra set.  This has two problems.  1)  Maybe our state is better
so we should take the intersection of both comparisons to get the most
accurate comparison.  2)  It over writes the implications by doing an
unnecessary set_states() which doesn't change anything.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>