comparison/core/hooks: pass other pointer to pre_merge_hooks
In the pre merge hooks, we pass in an sm but it's the sm from the other
side of the merge, not the current stree.  This caused a problem in
smatch_comparison.c because I assumed it was from the cur_stree.  This
lead to a lot of nonsense bugs.

My fix is to pass in both sm states.  I have a few places which look at
the other state so we can't just pass in the current sm state.  (You can't
look up the other state, because that stree isn't available).

Most places don't care about the current state.  The most common thing is
to just test if we're in an impossible path and mark the path as ok if so.
There were a few places which looked up the current sm_state so now that
code is a little bit simpler.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>