comparison, states: fix recursion in pre_merge_hooks()
The problem is that when we we're doing the get_implied_rl() it can lead
to a bunch of stree manipulation and then it means we start doing another
pre_merge recursively.

Really I want to do a more permanent fix so that we don't push fake cur
strees in smatch_math.c when we're doing a pre_merge.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>