typeof: examine it at show-time
Unless an explicit call to examine_pointer_target() or
get_base_type() is made, the base type of pointers are
*not* examined via the usual recursive examine_symbol_type().
That means that it is possible to call show_typename()
on a non-fully examined type which is wrong (for example,
because SYM_TYPEOFs may not be converted).

So, call examine_pointer_target() on pointers when trying
to display them.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>