typeof: add a test for unexamined typeof
The base type of pointers are not examined when the pointer is.
It needs to be done later when looked at.

This may be a problem when show_typename() is used on a pointer
which has not yet been 'deep-examined' and, for example, has a
SYM_TYPEOF as its base type.

Add a test case showing the problem.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>