flow: fix a crashing bug in parsing inlines
The problem code here looks like this:

static void target_append_str(char **str, const char *txt)
{
        char *prev = *str;

        *str = *str ? kasprintf(GFP_ATOMIC, "%s,%s", *str, txt) :
                kstrdup(txt, GFP_ATOMIC);
        kfree(prev);
}

The problem is when we first encounter the assignment:

	*str = *str ? kasprintf(GFP_ATOMIC, "%s,%s", *str, txt) :
		kstrdup(txt, GFP_ATOMIC);

At first *str is NULL and it's a parameter so Smatch decides treats it
as a known value (when we're parsing this inline).  The strip_expr() will
just return kstrdup() instead of the conditional.

But then after we do the assignment, the *str is not known so strip_expr()
will return the conditional.  The checks are expecting a function call so
they crash.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>