cgcc: remove _STRING_ARCH_unaligned
As far as I can see, the macro '_STRING_ARCH_unaligned' may
be defined by glibc, not gcc. So, there is no reason for
cgcc to define it.

Worse, cgcc defines it to '1' while on some platorms (at least the
few ppc64 I've access to) the installed glibc defines it to '0'.

So, do not let cgcc define this macro.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>