arch: char32_t should be the same as uint32_t, not uint
When the predefine for char32_t was added, it was made to
correspond 'unsigned int' with the commit message saying
some archs use 'unsigned long'. In fact, it appears that
char32_t is always uint32_t (on the archs & OSes I'm using
to look at this).

So, simply predefine __CHAR32_TYPE__ like uint32_t is.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>