add tests for function attributes
Function attributes need to be parsed differently than
the usual specifiers. For example, in code like:
	#define __noreturn __attribute__((noreturn))
	__noreturn void foo(int a);
the __noreturn attribute should apply to the function type,
while a specifier like 'const' would apply to its return type.
It's even more clear when function pointers are involved:
	__noreturn void (*fptr)(void);
here too, the attribute should be applied to the function type,
not the its return type, nor to the declared pointer type.

Add some testcases to cover some of the situations concerning
the parsing of these function pointers.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>