fix assignment check with function attribute
It's OK to assign a non-qualified type to the corresponding
const- or -volatile-qualified type (same for 'restrict').
For modifiers like __pure or __noreturn, it's the opposite:
it's OK to assign the qualified version to the corresponding
non-qualified type.

Allow this by using type_difference() with the appropriate mask
when checking assignments.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>