let function definition inherit prototype attributes
It's common to declare a function with the attribute
'pure' or 'noreturn' and to omit the attribute in the
function definition. It makes somehow sense since the
information conveyed by these attributes are destined
to the function users not the function itself.

So, when checking declaration/definition, let the
current symbol inherit any function attributes present
in previous declarations.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>