db: count states better in split_on_bool_sm()
The split_on_bool_sm() is used for boolean parameters and also for NULL
vs non-NULL parameters.  Boolean parameters are strictly "0-1" range.  I
maybe would have expected NULL vs non-NULL to be "0,4096-ptr_max" but it
can actually be a lot of different values.  The split_on_bool_sm() will
create a separate return for each of those different values.

Since it creates only 2 returns for boolean parameters and that means we can
allow more states and still try to split up the return.  But for the
NULL vs non-NULL, I left the code as is so it say "states * possible <= 2000"
is fine but more than 2000 is too much.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>