address: push latest code for get_address()
I'm sorry, I started to re-write this code but then I got distracted.
I've been running this code for a while.  I have been wanting to push it
for a while.  But I can't remember what problem I was trying to solve and
I'm pretty sure I didn't finish because I kept on finding that it was
worse and worse the deeper I looked.

Still, I just want to get rid of this diff from my local repository so I
am going to push it.  :(

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>