parsed_conditions: introduce a new module to handle already parsed conditions
One thing that we sometimes have is:

	return foo() ? ({ large macro; }) : ({ complicated thing; });

The way that this works is that the smatch_db.c parses the condition and
gets the states.  This works well enough when the condition is a variable
or a comparison, but it doesn't work well when it is a function.

The problem why it doesn't work is because functions don't have
implications tied to them.  And if the complicated thing; is an assignment
or has side effects, those states are lost.

The fix is to just store a link to the true and false strees for the
condition.  The stree is tied to the expression pointer for the condition.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>