locking: re-write to use the cross function database
In theory, it should have been straight forward to fix the locking check
to use the standard name format for locks and a few functions to store and
select locking information from the database.  But what happened instead
was that I started a re-write years ago and then just went crazy re-writing
stuff.

It would be a lot of work to split the changes into one thing per patch and
write sensible changelogs.  Emotionally, it is more satisfying to just
forget the original locking check existed and pretend this is an entirely
new thing.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>