fix expansion of initializer (mismatching size)
Currently, the expansion of constant initializers is done
whenever the offset in the initializer match the one
we're expanding.

However, it's not correct to do this expansion if their
size doesn't match since in this case the value of one
doesn't represent the value of the other.

Fix this by adding a check for the size.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>