fix expansion of initializer (mismatching type)
Currently, the expansion of constant initializers is done
whenever the offset in the initializer match the one
being expanded.

However, it's not correct to do this expansion of an
integer with the initializer for a float and vice-versa.

Fix this by adding the corresponding tests to the other
tests of the value.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>