fix addressability marking in evaluate_addressof()
mark_addressable() is used to track if a symbol has its
address taken but does not take in account the fact that
a symbol can be accessed via one of its subfields.
A failure occurs in case like:
	struct { int a; } s = { 3 };
	...
	def(&s.a);
	return s.a;

where 's' is not marked as being addressable and so the
the initializer will be expanded and the return expression
will always be replaced by 3, while def() can redefine it.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>