fix simplify_loads() when doing type punning
When doing loads simplification for a location where
floats & integers are mixed, loads are systematically
replaced with the value of their dominating memop (this
checks if the corresponding write or load overlaps).

However, this must not be done if the involved operations
are doing some form of integer/float type punning.

Fix this by refusing to convert load of an integer by a
previous float value or the opposite.

Note: another way to describe this problem would be to say
      that floats need to have their own memory operations:
          OP_FSTORE & OP_FLOAD
      or that instructions need to have some form of 'machine type'
      in addition of the size (like clang's i32/f32, ...).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>