add testcase for addressability of 'complex' symbols
Once a symbol has its address taken, a lot of simplifications
must be avoided because the symbol can now be modified via
a pointer.

This is currently done but the symbol addressability
does not take in account the fact that a symbol can be
accessed via one of its subfields.

Add a testcase to illustrate this.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>