fix cost of dereference of symbols with complex type
Currently, in expand_dereference(), the dereference of a symbol with
a complex type is considered as costing as high as a non-symbol
because it's not recognised it's a symbol.

However, both cases should have exactly the same cost since they
address calculation amounts to 'symbol + offset'.

So, instead of taking in account a single level of
	symbol + offset
let's use a loop for this in order to handle
	symbol [+ offset]*

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>