degenerated arrays & functions are addressable too
Symbols which have their address taken (with the 'addressof'
operator: &) are marked as such (with the modifier MOD_ADDRESSABLE).
But degenerated arrays and functions have their address implicitly
taken.

MOD_ADDRESSABLE is used to prevent to replace a symbol dereference
nto the value used to initialize to it. For example, in code like:
	static int foo(void)
	{
		int x[2] = { 1, 2 };
		return x[1];
	}

the return expression can be replaced by 2. This is not the case
case if the array is first passed in a function call, like here:
	extern void def(void *, unsigned int);
	static int bar(void)
	{
		int x[2] = { 1, 2 };
		def(x, sizeof(x));
		return x[1];
	}

Fix this by marking degenerated arrays (and functions) as also
being addressable.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>