container_of: swap variable names
We're parsing "foo->bar(foo);" so the argument is the container and the
function is the expression.  Passing the container is the common case, but
even if we have "foo->bar(foo->something);" where neither is a container,
the function is never a container.

This code works correctly but the names were reversed so it was confusing.
This patch shouldn't change runtime.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>