container_of: fix how container of star information is recorded
The container of information looks something like "*(-16+4)" or "-16+4".
The difference is supposed to be the first one is foo->fn(foo->bar) and
the next one is "foo->fn(&foo->bar);".

In this code the expression should always be a function pointer so if we
remove the &func and func, it should normally be safe.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>