db: improve split_success_fail
I bumped the number of states allowed.  This is a bit cheating and it
turned out to not be effective for long, but it didn't hurt anything so
we may as well.

Then the other thing is that a lot of times the DB isn't fully filled out
and we don't know that error codes are nicely in the normal error code
range of (-4095)-(-1) we just know that it's non-zero.  So this should
be allowed as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>