typeof: do not let classify_type() do its own SYM_TYPEOF expansion
SYM_TYPEOFs are expanded at examination time.
However, classify_type() does its own expansion of SYM_TYPEOFs.
Worse, it does this differently (address space & noderef are
not removed)..

So, to enforce the same expansion, also use examine_symbol_type()
to do the expansion in classify_type().

Note: it's not sure that it's currently possible to have
      SYM_TYPEOFs to expand in classify_type().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>