mark strings as examined & evaluated
evaluate_string() leaves the strings it creates as unexamined
& unevaluated. More exactly, they are examined and evaluated
(they have correct size & type) but not marked as such.

This doesn't seem to really matter but shows up when auditing
if classify_type() is always used on examined symbols.

So, mark the strings as examined and evaluated since their
size & type are known.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>