arch: move arch-specificities to their own files
lib.c and even more so, target.c, become cluttered by the
arch specific type initialization.
It would be better to move this to arch-specific files,
move the generics target related helpers to target.c
and have sparse_initialize() to just call these helpers.

For doing this:
* introduce a struct to hold the configurations for each arch,
* move the arch-specific type initialization to separate files,
* make target.c generic.

Also change the default types to LP64.

Note: this is the first step to better handle other arch
      specificities like the predefines or the handling
      of some options.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>